♻️ Embed fonts and images on demand
This commit defers the embedding of fonts and images in a PDF page. The
functions `getPageFont` and `getPageImage`, that are used to obtain a
reference to a font or an image object in the page, now also *add* the
font or the image if it is not already registered. Consequently, these
functions are renamed to `addPageFont` and `addPageImage`.

As an effect, images defined in the document definition are now only
included in the PDF when they are actually used in the document. For
fonts, this was already the case because the font subsetting kept track
of the glyphs being used.