✨ Support auto width for blocks
This commit adds support for shrinking blocks to the minimal width
required to display their content. To do so, the `width` attribute of a
`Block` now supports the value `auto` on all block types.

To compute the minimal required width, the `auto` width is passed down
to children that don't have a fixed width set.