🔧 Decouple Prettier from ESLint
We used to run Prettier as a part of ESLint using a plugin. However,
this practice is not recommended by Prettier [1] as it is slower and
adds an unnecessary layer of indirection. This commit removes the
`eslint-plugin-prettier` package and configures auto-formatting with
Prettier in VSCode instead. The formatting is adjusted in some files.

[1]: https://prettier.io/docs/en/integrating-with-linters