📝 Provide accessible examples
The example folder contained only a single example named `sample.js`.
This file was used to check certain conditions during development, but
was not comprehensible to users. To make the examples more accessible,
this commit replaces this file with multiple individual examples for
different topics.

A README file is added to the examples folder to explain how to run the
examples. A script is provided to download the required fonts. This is
clearer than extracting TTF fonts from WOFF files in npm packages. It is
also closer to an actual usage of the library.

The `npm start` magic is removed. All runtime environments support a
`--watch` flag that can be used to re-create the PDF files when the
source files change.