♻️ Introduce ImageLoader
In preparation for loading images on demand, this commit introduces a
new type `ImageLoader` to encapsulate the responsibility for loading
image data. This type is intended to become a public interface that
allows application code to load images on demand.

Instead of a string, the image to be loaded is now identified by an
`ImageSelector` that may also include the width and height of the target
area. This will allows selecting images that fit in size.