✨ Support fontWeight and fontStyle
The text attributes `bold` and `italic` only allow the selection of a
single alternative font weight and font style.

This commit introduces the text attributes `fontWeight` and `fontStyle`
as replacement for `bold` and `italic`. These new attributes correspond
to the CSS properties [font-weight] and [font-style]. This enables the
use of different weights of the same font family.

The attributes `bold` and `italic` are still supported, but they are
deprecated and will be removed in a future release.

[font-weight]: https://developer.mozilla.org/en-US/docs/Web/CSS/font-weight
[font-style]: https://developer.mozilla.org/en-US/docs/Web/CSS/font-style