✨ Load images from file system
Images need to be registered with the global `images` attribute. When
loading images from the file system, this shouldn't be necessary.

This commit allows selecting images by their file name, similar to the
`src` attribute in HTML `<img>` tags. When the name given in the `image`
attribute of an image block is not a registered name, it is now
interpreted as a file name and the image is loaded from the file system.
Relative paths are resolved relative to the current working directory.

The optional `format` attribute of an image definition is removed. The
image format is now auto-detected from the file content.

The `ImageStore` now caches the loaded images and errors from the image
loader. This prevents loading the same image again when it is used
multiple times.