🔧 Use esbuild, add npmignore
The TypeScript compiler creates a single JavaScript file for each
TypeScript module, and it insists on importing `.ts` files as `.js`.
[esbuild] creates a single bundled `index.js` which may be more
efficient to load than many individual files. It also supports `.ts`
imports.

This commit uses [esbuild] for building the package and uses the
TypeScript only to create type definitions.

An `.npmignore` file is added to exclude all files except the ones we
want to publish.

[esbuild]: https://esbuild.github.io/