Replace the term "attribute" with "property"
The API documentation used the term "attribute" to refer to the
properties of text, blocks, and graphics. However, since these
attributes are actually set as JavaScript object _properties_, the term
"property" is more appropriate. Furthermore, many of the properties in
this library are based on CSS _properties_ such as `line-height` or
`font-weight`. SVG only uses the term "attribute" because it inherited
it from XML.

This commit replaces the term "attribute" with "property" in the API
documentation. The types `TextAttrs`, `BlockAttrs`, `InfoAttrs`, and
`CustomInfoAttrs` are deprecated in favor of new properties `TextProps`,
`BlockProps`, `InfoProps`, and `CustomInfoProps`.