🔧 Only include public API types in package
The package should not contains any internal type definitions. This
commit excludes any `*.d.ts` files other than the `index.d.ts` and the
`d.ts` files in the `api` folder from the package. All API is contained
in this folder.

The `stripInternal` option is added to the `tsconfig.json` to allow
marking any internal types or functions in this folder with `@internal`.