🗑️ Replace type Text with TextSpan
The type `Text` had a recursive definition that included an array of
itself (`Text[]`). This created an unnecessary complex type that may
even cause problems for the type checker. It also allowed for the
unintended edge case of nested arrays (e.g. `[['foo']]`).

This commit and introduces a new, simpler type `TextSpan` that combines
a `text` property with optional style properties. The `text` property
also accepts nested `TextSpan` arrays. This results in a clearer and
almost identical interface except for the irregular edge case of nested
arrays.

The type `Text` is marked as deprecated.