⬆️ Upgrade ESLint to v9, adjust config, fix linter issues
This commit upgrades ESLint to v9. ESLint plugins are also updated to
the latest versions. The `.eslintrc` and `eslintignore` files are
replaced with a single `eslint.config.js` file that follows the new
ESLint configuration format [1]. Some useful rules are added to the
configuration and the resulting linter issues are fixed.

The ESLint option `--format visualstudio` has been moved to a separate
npm package `eslint-formatter-visualstudio`. We leave it out, as the
convenience of the output format does not justify the additional
dependency.

[1]: https://eslint.org/docs/latest/use/configure/migration-guide