🔧 Type-check tests in CI pipeline
By excluding test files in the `tsconfig.json` file, the TypeScript
compiler did not check these files. Consequently, issues in the tests
only showed up in the editor, but not in the CI build.

This commit removes the patterns for test files from the `exclude`
section in the `tsconfig.json` file. To ensure that only relevant files
end up in the `dist` directory, we let the TypeScript compiler write to
a `build/` directory and then copy the necessary files to `dist/`.