♻️ Require await in async functions
This commit enables the ESLint rule `require-await` which requires the
use of `await` in async functions and fixes issues that have been
uncovered by this rule. This rule helps to spot unintentional and
leftover `async` keywords.

The `loadFont` function in `FontLoader` was unnecessarily asynchronous
and has been changed to be synchronous. This change does not affect the
API.