⬆️ Require Node 20, npm 10
Node.js version 20 is the current LTS version. Node.js 18 will be
end-of-life in April 2025. This commit updates the minimum required
Node.js version to 20.

Node 20 supports ESM out of the box. The WebCrypto API is available in
the global scope. The `--experimental-global-webcrypto` flag is no
longer needed to run the examples.

The manual exposure of the API in the global scope has been removed from
the tests.