Set error cause property
Since EcmaScript 2022, `Error` objects support a `cause` property,
allowing for tracking the root cause of an error.

This commit set this property where appropriate.

Since vitest's `toThrowError()` does not support checking the `cause` of
an error [1], we use utility functions to catch and extract errors
thrown by a function. This allows checking for individual properties of
an error.

[1]: https://github.com/vitest-dev/vitest/issues/5697