♻️ Merge FontLoader with FontStore
Fonts aren't actually _loaded_, as the data is already part of the font
definition. The class `FontLoader` actually _selected_ the font from the
set of registered fonts. The separation between `FontLoader` and
`FontStore` is not necessary.

This commit merges the two classes into `FontStore`. The interface is
kept async in preparation for loading fonts on demand.