🐛 Fix error with custom font name, test with read fonts
Setting a custom name on a font registration lead to an error in the
`FontStore`. This was caused by passing the custom name to fontkit's
`create` method which only supports this on certain font types.

This commit fixes the error and adds a test using real font files to
ensure the registration works in practice.