✨ Support including images by URL
Images used to be registered in the document definition. In v0.5.4,
support for loading images directly from the file system was added.
However, this feature gives the library access to all files on the
system, which may be a security concern.

This commit adds support for including images by a URL. The `image`
property of an image block now supports `data:`, `file:`, and `http(s):`
URLs. File names are interpreted as relative to a resource root that
must be set by the `setResourceRoot()` method on the `PdfMaker` class.
Compatibility with the previous API is maintained, but the `images`
property in a document definition is deprecated.

A new `fs` module is introduced to read files from the file system.
Since the `node:` modules are not available in browser environments,
they are loaded using dynamic imports. This required to raise the
minimal runtime version to ES2022. The build has been updated to target
the browser platform and ES2022. This ensures that the library can be
used in the browser, but still can access the file system in Node.