🐛 Fix TypeError for color on nested blocks
Reading a `color` property on a row or column block with a nested text
block lead to a TypeError:

> TypeError: Invalid value for "definition/content/0/rows/0/color":
> Expected valid color, got: {type: 'RGB', red: 0, green: 0.3, blue:
> 0.5}

This happened because the `readColor()` function transforms the input
and did not accept its own result type as input. For nested blocks, the
inherited `color` property is parsed twice if not overwritten.

As a quick fix, this commit changes the `readColor()` function to accept
its own result type, i.e. an object with type `RGB`.

Fixes https://github.com/eclipsesource/pdf-maker/issues/95