🩹 Fix type error for keywords in metadata
The type of the `info` property in the document definition combines the
`InfoProps` and `CustomInfoProps` types. This introduces a type conflict
as the type of the `keywords` property (`string[]`) conflicts with the
type of the values in `CustomInfoProps` (`string`). As a result, an
assignment of the `keywords` property results in a type error.

To fix this problem, we enforce the `XX` prefix for custom keys in the
`info` property, following the recommendation. This could be considered
a breaking change, however, it will only affect custom info attributes
without the recommended `XX` prefix and the resulting type error can
easily be suppressed.