🐛 Fix error when reusing fonts or images
The internal `Font` and `Image` objects are kept in global stores in the
PdfMaker context. During the generation of a document, these objects got
PDFRefs attached that outlived the document and caused errors during the
generation of subsequent documents.

The solution is to attach these references on the PDFDocument instead.